--- functions/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "commonjs",
-    "noImplicitAny": true,
-    "removeComments": true,
-    "preserveConstEnums": true,
-    "forceConsistentCasingInFileNames": true,
-    "outDir": "./js"
-  },
-  "include": ["ts/*"]
-  // "files": ["ts/hello.ts"],
-}
