+++ package.json
@@ -20,12 +20,12 @@
     "typescript": "~3.7.2"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "start:lambda": "netlify-lambda serve functions/js",
+    "start": "concurrently \"yarn start:client\" \"yarn start:lambda\"",
+    "start:client": "react-scripts start",
+    "start:lambda": "netlify-lambda serve functions",
     "build": "yarn build:client && yarn build:lambda",
     "build:client": "react-scripts build",
     "build:lambda": "netlify-lambda build functions",
-    "build:lambda-compile": "tsc ./functions/ts/hello.ts",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
     "format": "prettier --write .",
@@ -49,6 +49,7 @@
   "devDependencies": {
     "@types/aws-lambda": "^8.10.61",
     "bestzip": "^2.1.6",
+    "concurrently": "^5.3.0",
     "husky": "^4.2.5",
     "lint-staged": "^10.2.11",
     "netlify-lambda": "^2.0.1",
